generate PublicObjects
I could do it automatically in hyzzy, but that would force the player to
define all the objects in Objects.hs and to use regenPublicObjects.sh; I
want to make things easier in the simple case, but I don't want to be a
burden if the user prefers to e.g. split the Objects file into several
modules, or define the room-specific objects in the appropriate room.